+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 ### Added
 
 - ✨(backend) include ancestors accesses on document accesses list view # 846
-- ✨(backend) add ancestors links definitions to document abilities #846
+- ✨(backend) add ancestors links reach and role to document API #846
 - ✨(frontend) add customization for translations #857
 - ✨(frontend) Duplicate a doc #1078
 - 📝(project) add troubleshoot doc #1066